refactor: add stubs
